User Uploads (Non-Functional)
'm having trouble getting the JSON string converted back and mapped to anything usable. I'm going to push what I have, but it is not functional. I need to figure out a way to iterate over it the upload data and to get the count of upload entries.

I have used this and similar-looking sources and can't figure it out.
https://stackoverflow.com/questions/40826918/how-to-map-json-to-an-object-with-a-different-structure